fix: payout計算をint(round(...))から10円単位切り捨てに修正
直前のf14c344でint()切り捨てをint(round())に変更したが、
浮動小数点誤差やオッズ値の精度次第で1の位が0にならない
払戻額が残る可能性があった（例: 2,709円など）。競輪の公式
払戻金は控除率計算後に必ず10円単位に切り捨てられるため、
round()で誤差を吸収した上で10円単位に切り捨てる処理へ統一。

- src/evaluation/backtest_wt.py: notify_results_wt.py本番経路
- scripts/sync_payout_to_vps.py: VPS payout同期スクリプト
- scripts/save_model_eval.py: model_evaluation配当計算
- scripts/backfill_picks_history_wt.py: picks_historyバックフィル

+++ scripts/backfill_picks_history_wt.py
@@ -232,7 +232,8 @@ def backfill_period(
             for t in combo_frames:
                 cs = frozenset({pivot1, pivot2, t})
                 if cs == top3_set:
-                    payout = int(round(combo_odds.get(t, 0) * 100))
+                    # 公式払戻金は10円単位に切り捨て
+                    payout = round(combo_odds.get(t, 0) * 100) // 10 * 10
                     hit = True
                     break
             history.append((race_date, store_key, rank, pred, n_combos,
